refactor(logging): split aigear-logs logic into deploy and infra modules
Move run discovery, cache, and gcloud logging read out of the CLI entry point so gcp_logs.py only handles argument parsing and user interaction, matching the scheduler CLI layering.
